Extracted example bots into separate files
