+++ doom-player/src/client/app.tsx
@@ -2,1849 +2,15 @@ import { useCallback, useEffect, useRef, useState } from "react";
 import CodeMirror from "@uiw/react-codemirror";
 import { javascript } from "@codemirror/lang-javascript";
 
-// Simple default bot: walks forward for ten ticks, logging HP and
-// screen each step. Designed to be the smallest useful demonstration
-// of the codemode API; the "Examples" picker offers more involved
-// bots for users to graduate to.
-const SIMPLE_BOT = `// A minimal codemode bot. The game is already past the menus when
-// this code starts, so we can read state and press keys right away.
-//
-// API (all async, always 'await'):
-//   await bot.getState()            // engine snapshot (hud, screen, ...)
-//   await bot.press(key, holdMs?)   // key tap or hold
-//   await bot.sleep(ms)             // pause between actions
-//   await bot.log(...args)          // streamed live to the log pane
-
-for (let i = 0; i < 10; i++) {
-  const s = await bot.getState();
-  await bot.log("tick", i, "screen:", s.screen, "hp:", s.hud.health);
-  await bot.press("up", 250); // walk forward for 250ms
-  await bot.sleep(50);
-}
-
-return "walked 10 steps";
-`;
-
-// More involved example: handles non-playing screens, fires at any
-// enemy directly ahead, and walks forward otherwise. Closer in spirit
-// to what a real bot looks like.
-const COMBAT_BOT = `// Walk forward and shoot enemies in the centre of the FOV.
-for (let i = 0; i < 40; i++) {
-  const s = await bot.getState();
-  await bot.log("tick", i, "screen:", s.screen, "hp:", s.hud.health,
-    "enemies:", s.enemies_visible.length);
-
-  // Press enter on menu / intermission / finale to advance.
-  if (s.screen !== "playing") {
-    await bot.press("enter");
-    await bot.sleep(150);
-    continue;
-  }
-
-  const centred = s.enemies_visible.find((e) => e.bearing === "center");
-  if (centred) {
-    await bot.log("firing at", centred.type);
-    await bot.press("fire", 200);
-  }
-  await bot.press("up", 250);
-  await bot.sleep(50);
-}
-
-return "combat run complete";
-`;
-
-// Full deterministic auto-player. Faithful port of the original
-// host-side /play loop into a single codemode bot. Priority-ordered
-// playing policy: fire centred enemies → turn toward off-centre
-// enemies → unstick → open doors / switches → take exits → grab
-// useful pickups → navigate toward the deepest open ray. Designed
-// to clear a level autonomously without an LLM in the loop.
-const AUTOPLAY_BOT = `// Deterministic Doom auto-player.
-//
-// Priority each tick (highest first):
-//   1. Centre-FOV enemy           -> fire
-//   2. Off-centre enemy           -> turn toward it
-//   3. Hard wedge (>=6 stationary ticks) -> 700ms right turn
-//   4. Door / switch ahead, close -> press use, step forward
-//   4b. Door / switch off-centre  -> turn toward it (approach)
-//   5. Exit at <=200 units        -> press use, step forward
-//   6. Useful pickup              -> turn toward / approach it (sticky)
-//   7. Soft wedge (>=3 stationary ticks) -> 350ms right turn
-//   8. Navigate                   -> follow deepest open ray
-//
-// Tuning constants -- tweak these to change behaviour.
-const MAX_TICKS = 400;
-const TICK_MS = 250;
-
-// Wedge detection: ticks before nudging / spinning when motion stalls.
-const SOFT_WEDGE_TICKS = 3;
-const HARD_WEDGE_TICKS = 6;
-// Squared distance below which a frame counts as "didn't move"
-// (player radius is 16; 32^2 filters out grazing micro-motion).
-const WEDGE_EPSILON_SQ = 32 * 32;
-
-// Cooldown ticks after pressing use on a door / after an unwedge
-// nudge, to avoid spamming use while the door animates open or
-// re-targeting the same unreachable pickup we just spun away from.
-const DOOR_COOLDOWN = 4;
-const UNWEDGE_COOLDOWN = 3;
-
-// Sticky pickup target lifetime (ticks before we give up on the
-// type we locked onto and let other policies run).
-const PICKUP_LIFETIME = 20;
-
-// Category priorities. Lower = preferred.
-const PICKUP_PRIORITY = {
-  key: 0,
-  weapon: 1,
-  powerup: 2,
-  armor: 3,
-  health: 4,
-  ammo: 5,
-};
-
-// --- Pure helpers ----------------------------------------------------
-
-function turnKeyForBearing(bearing) {
-  if (bearing === "far_left" || bearing === "left") return "left";
-  if (bearing === "right" || bearing === "far_right") return "right";
-  return null;
-}
-
-function bearingScore(bearing) {
-  if (bearing === "center") return 0;
-  if (bearing === "left" || bearing === "right") return 1;
-  return 2;
-}
-
-function distanceScore(distance) {
-  if (distance === "near") return 0;
-  if (distance === "mid") return 1;
-  if (distance === "far") return 2;
-  return 3;
-}
-
-function pickClosestEnemy(enemies) {
-  const sorted = [...enemies].sort((a, b) => {
-    const ab = bearingScore(a.bearing) - bearingScore(b.bearing);
-    if (ab !== 0) return ab;
-    return distanceScore(a.distance) - distanceScore(b.distance);
-  });
-  return sorted[0];
-}
-
-function bearingTurnMs(bearing) {
-  if (bearing === "left" || bearing === "right") return 140;
-  if (bearing === "far_left" || bearing === "far_right") return 320;
-  return 0;
-}
-
-function pickupIsUseful(thing, hud) {
-  if (!(thing.category in PICKUP_PRIORITY)) return false;
-  if (thing.category === "health") {
-    if (thing.type === "stimpack" || thing.type === "medikit") {
-      return hud.health < 100;
-    }
-    return hud.health < 200;
-  }
-  if (thing.category === "armor") {
-    if (thing.type === "green_armor") return hud.armor < 100;
-    if (thing.type === "blue_armor") return hud.armor < 200;
-    return hud.armor < 200;
-  }
-  return true;
-}
-
-function pickPickup(things, hud) {
-  const useful = things.filter((t) => pickupIsUseful(t, hud));
-  if (useful.length === 0) return undefined;
-  useful.sort((a, b) => {
-    const ap = PICKUP_PRIORITY[a.category] ?? 99;
-    const bp = PICKUP_PRIORITY[b.category] ?? 99;
-    if (ap !== bp) return ap - bp;
-    return a.distance - b.distance;
-  });
-  return useful[0];
-}
-
-// --- Action helpers --------------------------------------------------
-
-async function approach(bearing_deg, distance) {
-  const abs = Math.abs(bearing_deg);
-  if (abs > 10) {
-    const key = bearing_deg < 0 ? "left" : "right";
-    const ms = Math.min(500, Math.max(60, abs * 5));
-    await bot.press(key, ms);
-    return "turn:" + key;
-  }
-  const ms = Math.min(800, Math.max(150, distance * 2));
-  await bot.press("forward", ms);
-  return "forward";
-}
-
-async function navigate(state) {
-  const rays = state.raycasts;
-  if (rays.length === 0) {
-    await bot.press("forward", 200);
-    return "blind_forward";
-  }
-  const ranked = [...rays].sort((a, b) => {
-    const w = (r) => (r.hit === "open" ? r.distance + 2000 : r.distance);
-    return w(b) - w(a);
-  });
-  const best = ranked[0];
-  const centre = rays
-    .filter((r) => Math.abs(r.bearing_deg) <= 20)
-    .sort((a, b) => b.distance - a.distance)[0];
-  if (centre && centre.distance >= 200) {
-    await bot.press("forward", 400);
-    return "navigate_forward(" + centre.distance + ")";
-  }
-  return await approach(best.bearing_deg, best.distance);
-}
-
-// --- Main loop -------------------------------------------------------
-
-const actions = {};
-const bump = (name) => {
-  actions[name] = (actions[name] ?? 0) + 1;
-};
-
-// Per-loop mutable state (replaces the DoomPlayer private fields).
-let lastPos = null;
-let stuckTicks = 0;
-let doorCooldown = 0;
-let unwedgeCooldown = 0;
-let pickupTarget = null;
-
-function updateWedge(pose) {
-  if (!pose) {
-    stuckTicks = 0;
-    return;
-  }
-  if (lastPos === null) {
-    lastPos = { x: pose.x, y: pose.y };
-    stuckTicks = 0;
-    return;
-  }
-  const dx = pose.x - lastPos.x;
-  const dy = pose.y - lastPos.y;
-  const moved = dx * dx + dy * dy;
-  lastPos = { x: pose.x, y: pose.y };
-  if (moved < WEDGE_EPSILON_SQ) {
-    stuckTicks++;
-  } else {
-    stuckTicks = 0;
-  }
-}
-
-async function playStep(state) {
-  if (doorCooldown > 0) doorCooldown--;
-  if (unwedgeCooldown > 0) unwedgeCooldown--;
-  updateWedge(state.player);
-  const wedgedSoft = stuckTicks >= SOFT_WEDGE_TICKS;
-  const wedgedHard = stuckTicks >= HARD_WEDGE_TICKS;
-
-  if (pickupTarget) {
-    const stillVisible = state.things_visible.some(
-      (t) => t.type === pickupTarget.type,
-    );
-    if (!stillVisible) {
-      pickupTarget = null;
-    } else {
-      pickupTarget.ticksRemaining--;
-      if (pickupTarget.ticksRemaining <= 0) pickupTarget = null;
-    }
-  }
-
-  // 1. Centre enemy -> fire.
-  const centre = state.enemies_visible.find((e) => e.bearing === "center");
-  if (centre) {
-    await bot.press("fire", 120);
-    pickupTarget = null;
-    return "fire:" + centre.type;
-  }
-
-  // 2. Off-centre enemy -> turn toward.
-  const off = pickClosestEnemy(state.enemies_visible);
-  if (off) {
-    const key = turnKeyForBearing(off.bearing);
-    if (key) {
-      await bot.press(key, bearingTurnMs(off.bearing));
-      pickupTarget = null;
-      return "aim_enemy:" + off.bearing;
-    }
-  }
-
-  // 3. Hard wedge -> big right turn.
-  if (wedgedHard) {
-    await bot.press("right", 700);
-    stuckTicks = 0;
-    pickupTarget = null;
-    unwedgeCooldown = UNWEDGE_COOLDOWN;
-    return "unwedge_hard";
-  }
-
-  // 4. Door / switch handling. Doom's "use" key activates both:
-  //    - door lines (open / close / lock-and-unlock)
-  //    - switch lines (raise lifts, open remote doors, end-level
-  //      switches that aren't tagged as exits, etc.)
-  //    Both surface through raycasts as hit="door" or hit="switch".
-  //    Treat them the same: walk up, press use.
-  //
-  //    Pick the ray with the smallest off-centre bearing first; if
-  //    multiple usable lines are in view, the most-aligned one is
-  //    usually the one the level designer intended us to interact
-  //    with.
-  const usableRays = state.raycasts
-    .filter((r) => r.hit === "door" || r.hit === "switch")
-    .sort((a, b) => Math.abs(a.bearing_deg) - Math.abs(b.bearing_deg));
-  const usableRay = usableRays[0];
-  if (usableRay && doorCooldown === 0) {
-    const aligned = Math.abs(usableRay.bearing_deg) <= 15;
-    // 4a. Close + roughly centred -> press use, step through.
-    if (aligned && usableRay.distance <= 80) {
-      await bot.press("use", 80);
-      await bot.press("forward", 300);
-      doorCooldown = DOOR_COOLDOWN;
-      return "open_" + usableRay.hit + "(d=" + usableRay.distance + ")";
-    }
-    // 4b. Within reach but not yet aligned / close -> approach.
-    //     Skip while wedged so we don't fight the unstick logic.
-    if (!wedgedSoft && usableRay.distance <= 400) {
-      const action = await approach(usableRay.bearing_deg, usableRay.distance);
-      // Drop any pickup lock; doors/switches usually gate progress.
-      pickupTarget = null;
-      return "approach_" + usableRay.hit + "(" + action + ")";
-    }
-  }
-
-  // 5. Exit close ahead -> use, step forward.
-  const exitRay = state.raycasts.find(
-    (r) => r.hit === "exit" && r.distance <= 200,
-  );
-  if (exitRay) {
-    await bot.press("use", 80);
-    await bot.press("forward", 400);
-    return "exit_level(d=" + exitRay.distance + ")";
-  }
-
-  // 6. Pickup approach (sticky).
-  if (!wedgedSoft && unwedgeCooldown === 0) {
-    let target = null;
-    if (pickupTarget) {
-      const current = state.things_visible.find(
-        (t) => t.type === pickupTarget.type,
-      );
-      if (current && pickupIsUseful(current, state.hud)) {
-        target = current;
-      } else {
-        pickupTarget = null;
-      }
-    }
-    if (!target) {
-      target = pickPickup(state.things_visible, state.hud);
-      if (target) {
-        pickupTarget = { type: target.type, ticksRemaining: PICKUP_LIFETIME };
-      }
-    }
-    if (target) {
-      const action = await approach(target.bearing_deg, target.distance);
-      return "grab:" + target.type + "(" + action + ")";
-    }
-  }
-
-  // 7. Soft wedge -> shorter right turn.
-  if (wedgedSoft) {
-    await bot.press("right", 350);
-    pickupTarget = null;
-    return "unwedge_soft";
-  }
-
-  // 8. General navigation.
-  return await navigate(state);
-}
-
-let lastScreen = "unknown";
-let lastHp = -1;
-
-for (let tick = 1; tick <= MAX_TICKS; tick++) {
-  let state;
-  try {
-    state = await bot.getState();
-  } catch (err) {
-    // Engine may briefly fail get_state during screen transitions.
-    await bot.log("t=" + tick, "wait_no_state", String(err.message ?? err));
-    bump("wait_no_state");
-    await bot.sleep(TICK_MS);
-    continue;
-  }
-
-  lastScreen = state.screen;
-  lastHp = state.hud.health;
-
-  let action;
-  switch (state.screen) {
-    case "intermission":
-    case "finale":
-      await bot.press("enter");
-      action = "advance_intermission";
-      break;
-    case "dead":
-      await bot.press("use");
-      action = "respawn";
-      break;
-    case "automap":
-      await bot.press("tab");
-      action = "close_automap";
-      break;
-    case "playing":
-      action = await playStep(state);
-      break;
-    case "title":
-    case "demo":
-    case "menu":
-      // Shouldn't happen post-preroll, but be defensive: try enter.
-      await bot.press("enter");
-      action = "menu_enter";
-      break;
-    default:
-      action = "wait_unknown";
-      await bot.sleep(100);
-      break;
-  }
-  bump(action);
-
-  // Log only interesting events every tick; throttle navigation /
-  // wait actions to every 10 ticks so the stream stays scannable.
-  const interesting =
-    action.startsWith("fire") ||
-    action.startsWith("open_door") ||
-    action.startsWith("open_switch") ||
-    action.startsWith("approach_door") ||
-    action.startsWith("approach_switch") ||
-    action.startsWith("exit_level") ||
-    action.startsWith("grab:") ||
-    action.startsWith("aim_enemy") ||
-    action.startsWith("unwedge") ||
-    action === "respawn" ||
-    action === "advance_intermission" ||
-    action === "wait_no_state" ||
-    state.screen !== "playing";
-  if (interesting || tick % 10 === 0) {
-    const pos = state.player
-      ? " pos=(" + state.player.x + "," + state.player.y + ") ang=" + state.player.angle_deg
-      : "";
-    await bot.log(
-      "t=" + String(tick).padStart(3, " "),
-      "screen=" + state.screen,
-      "action=" + action,
-      "hp=" + state.hud.health,
-      "armor=" + state.hud.armor,
-      "ammo=" + state.hud.ammo,
-      "enemies=" + state.enemies_visible.length,
-      "things=" + state.things_visible.length + pos,
-    );
-  }
-
-  await bot.sleep(TICK_MS);
-}
-
-const summary = Object.entries(actions)
-  .sort(([, a], [, b]) => b - a)
-  .map(([n, c]) => n + "=" + c)
-  .join(" ");
-await bot.log("done; finalScreen=" + lastScreen, "finalHp=" + lastHp);
-await bot.log("actions:", summary);
-return { ticks: MAX_TICKS, finalScreen: lastScreen, finalHp: lastHp, actions };
-`;
-
-// One-shot diagnostic bot: dumps the first state snapshot and exits.
-// Useful for inspecting what fields the engine exposes without
-// writing a loop.
-// Vision-style exploration bot that doesn't use any LLM. Maintains
-// a pixel-quantised memory map of explored space using raycasts +
-// player position, then picks the next direction by looking for the
-// least-covered octant around the player. Dumps the memory canvas to
-// the debug-image panel every few macros so a human can watch the
-// map fill in.
-const EXPLORE_BOT = `// Memory-based exploration: paint what raycasts see, head where you
-// haven't been. No LLM — the engine's raycasts + position are enough.
-//
-// Memory canvas: a single high-resolution pixel buffer that lives for
-// the whole run. World units -> pixels via SCALE = 8 (so 1 pixel == 8
-// map units). At 800x800 pixels that covers a 6400x6400 unit area —
-// generous for Doom 1 / 2 level extents. Doom doesn't expose level
-// bounds via get_state (see AM_findMinMaxBoundaries in am_map.c) so we
-// pre-size and centre on the player's first observed position.
-//
-// Each tick we update three things:
-//   1. The 8 visibility rays the engine reports. Cells the ray passed
-//      through -> FLOOR; the hit cell -> WALL / DOOR / SWITCH / EXIT /
-//      THING. Higher-rank observations never get overwritten.
-//   2. Visible things (pickups, decor, enemies, barrels): painted at
-//      their world position and tracked in a sticky map. If a tracked
-//      thing close to the player disappears from things_visible we
-//      assume the player just picked it up (or shot it) and clear
-//      the pixel back to FLOOR.
-//   3. The player's own cell, stamped PLAYER.
-//
-// Direction picking divides the world around the player into 8
-// octants and picks the one whose lookahead area is least-explored,
-// penalising octants that are mostly wall.
-//
-// Debug dump: instead of sending the whole 800x800 canvas (huge PNG,
-// hard to scan), we clip a small window centred on the player and
-// log THAT. The crop is sized so 1 source pixel == 1 output pixel
-// (no scaling) — keeps detail crisp in the side panel.
-
-const SCALE = 8;                 // map units per memory pixel
-const SIZE  = 800;               // memory canvas edge, in pixels
-const CENTRE = SIZE / 2;
-
-const CROP = 256;                // edge of the clipped debug view
-const CROP_HALF = CROP / 2;
-
-const STEPS = 6;
-const TICKS_PER_MACRO = 12;
-const TICK_PAUSE_MS = 25;
-const LOG_EVERY = 2;             // log clipped view every N macros
-const TURN_TOLERANCE_DEG = 15;
-const DEBUG_TICKS = true;
-
-// Distance under which a previously-seen thing that has fallen out
-// of things_visible is presumed picked up (in map units).
-const PICKUP_RADIUS = 96;
-
-const INTEREST_PRIORITY = {
-  exit: 0,
-  switch: 1,
-  door: 2,
-  key: 3,
-};
-
-const OPPORTUNISTIC_PICKUPS = new Set(["weapon", "powerup", "armor", "health", "ammo"]);
-
-// Palette (RGBA bytes). UNSEEN must stay all-zero R/G/B so an
-// initially-zero buffer reads as "unexplored" without painting it.
-const COL = {
-  UNSEEN:  [  0,   0,   0, 255],
-  FLOOR:   [ 40,  40,  40, 255],
-  VISITED: [  0,  90, 180, 255],
-  RAY:     [ 60,  60, 100, 255],
-  WALL:    [220, 220, 220, 255],
-  DOOR:    [240, 200,  40, 255],
-  SWITCH:  [  0, 200, 200, 255],
-  EXIT:    [ 80, 240,  80, 255],
-  THING:   [220,  60, 220, 255],
-  ENEMY:   [255, 120,  40, 255],
-  PLAYER:  [255,  60,  60, 255],
-};
-
-const RANK = {
-  UNSEEN: -1, FLOOR: 0, RAY: 1, VISITED: 2, PLAYER: 3, THING: 4, ENEMY: 5,
-  SWITCH: 6, DOOR: 7, EXIT: 8, WALL: 9,
-};
-
-// Memory canvas. RGB starts at 0 (UNSEEN); alpha must be 255 so the
-// PNG renders opaque. One byte of rank per pixel runs alongside so
-// observations can never downgrade.
-const mem = new Uint8Array(SIZE * SIZE * 4);
-for (let i = 0; i < SIZE * SIZE; i++) mem[i * 4 + 3] = 255;
-const rank = new Int8Array(SIZE * SIZE);
-rank.fill(RANK.UNSEEN);
-
-// Sticky thing tracker, keyed by quantised world position. Each
-// entry stores { x, y, category, lastSeenMacro, lastSeenTick }.
-// We use a Map so deletions are O(1) when the player picks
-// something up.
-const things = new Map();
-const thingKey = (wx, wy) => Math.round(wx / 8) + "," + Math.round(wy / 8);
-
-let originX = null;
-let originY = null;
-
-function paint(px, py, name) {
-  if (px < 0 || py < 0 || px >= SIZE || py >= SIZE) return;
-  const idx = py * SIZE + px;
-  if (rank[idx] >= RANK[name]) return;
-  rank[idx] = RANK[name];
-  const col = COL[name];
-  const o = idx * 4;
-  mem[o] = col[0]; mem[o + 1] = col[1]; mem[o + 2] = col[2]; mem[o + 3] = col[3];
-}
-
-// Force-paint regardless of rank — used when an item is picked up so
-// THING/ENEMY pixels revert to FLOOR.
-function repaint(px, py, name) {
-  if (px < 0 || py < 0 || px >= SIZE || py >= SIZE) return;
-  const idx = py * SIZE + px;
-  rank[idx] = RANK[name];
-  const col = COL[name];
-  const o = idx * 4;
-  mem[o] = col[0]; mem[o + 1] = col[1]; mem[o + 2] = col[2]; mem[o + 3] = col[3];
-}
-
-function worldToPx(wx, wy) {
-  return {
-    px: Math.round(CENTRE + (wx - originX) / SCALE),
-    py: Math.round(CENTRE - (wy - originY) / SCALE),
-  };
-}
-
-function paintRay(playerX, playerY, worldBearingDeg, distance, hitKind) {
-  const rad = (worldBearingDeg * Math.PI) / 180;
-  const dx = Math.cos(rad);
-  const dy = Math.sin(rad);
-  const stepUnits = SCALE; // 1 pixel per step now
-  const steps = Math.max(1, Math.floor(distance / stepUnits));
-  for (let s = 1; s < steps; s++) {
-    const p = worldToPx(playerX + dx * s * stepUnits, playerY + dy * s * stepUnits);
-    paint(p.px, p.py, "FLOOR");
-  }
-  const hp = worldToPx(playerX + dx * distance, playerY + dy * distance);
-  const name = hitKind === "door" ? "DOOR"
-            : hitKind === "switch" ? "SWITCH"
-            : hitKind === "exit" ? "EXIT"
-            : hitKind === "thing" ? "THING"
-            : hitKind === "open" ? "FLOOR"
-            : "WALL";
-  paint(hp.px, hp.py, name);
-}
-
-// Convert a "thing" sighting (bearing+distance) into a world position
-// and record it in the sticky tracker.
-function recordThing(state, t, macro, tick) {
-  const p = state.player;
-  if (!p) return;
-  const worldBearing = p.angle_deg - t.bearing_deg;
-  const rad = (worldBearing * Math.PI) / 180;
-  const wx = p.x + Math.cos(rad) * t.distance;
-  const wy = p.y + Math.sin(rad) * t.distance;
-  const key = thingKey(wx, wy);
-  const cat = t.category || "thing";
-  const colName = cat === "enemy" || cat === "barrel" ? "ENEMY" : "THING";
-  things.set(key, { x: wx, y: wy, category: cat, colName, lastSeenMacro: macro, lastSeenTick: tick });
-  const px = worldToPx(wx, wy);
-  paint(px.px, px.py, colName);
-}
-
-// If a tracked thing is very close to the player and we DIDN'T see
-// it this tick, assume the player picked it up (or killed it) and
-// erase the pixel back to FLOOR. Avoids stale magenta dots dotting
-// the memory map after the player walked through.
-function sweepPickups(state, currentlyVisibleKeys) {
-  const p = state.player;
-  if (!p) return 0;
-  let removed = 0;
-  for (const [key, t] of things) {
-    if (currentlyVisibleKeys.has(key)) continue;
-    const d = Math.hypot(p.x - t.x, p.y - t.y);
-    if (d > PICKUP_RADIUS) continue;
-    // Out of sight + within pickup range == gone.
-    const px = worldToPx(t.x, t.y);
-    repaint(px.px, px.py, "FLOOR");
-    things.delete(key);
-    removed++;
-  }
-  return removed;
-}
-
-function recordObservation(state, macro, tick) {
-  const p = state.player;
-  if (!p) return;
-  if (originX === null) { originX = p.x; originY = p.y; }
-
-  // Rays first so floor coverage is laid down before things stamp
-  // over it.
-  for (const r of (state.raycasts || [])) {
-    const worldBearing = p.angle_deg - r.bearing_deg;
-    paintRay(p.x, p.y, worldBearing, r.distance, r.hit);
-  }
-
-  // Things: mark visible ones AND sweep ones we just walked through.
-  const visibleKeys = new Set();
-  for (const t of (state.things_visible || [])) {
-    const worldBearing = p.angle_deg - t.bearing_deg;
-    const rad = (worldBearing * Math.PI) / 180;
-    const wx = p.x + Math.cos(rad) * t.distance;
-    const wy = p.y + Math.sin(rad) * t.distance;
-    visibleKeys.add(thingKey(wx, wy));
-    recordThing(state, t, macro, tick);
-  }
-  sweepPickups(state, visibleKeys);
-
-  // Player cell + a small cross stamp for visibility.
-  const me = worldToPx(p.x, p.y);
-  paint(me.px - 1, me.py, "VISITED");
-  paint(me.px + 1, me.py, "VISITED");
-  paint(me.px, me.py - 1, "VISITED");
-  paint(me.px, me.py + 1, "VISITED");
-  paint(me.px, me.py, "PLAYER");
-}
-
-function angleDelta(a, c) {
-  let d = (a - c) % 360;
-  if (d > 180) d -= 360;
-  if (d <= -180) d += 360;
-  return d;
-}
-
-function normAngle(a) {
-  return ((a % 360) + 360) % 360;
-}
-
-function thingPriority(t) {
-  const cat = t.category || "thing";
-  if (!(cat in INTEREST_PRIORITY)) return null;
-  return INTEREST_PRIORITY[cat];
-}
-
-function opportunisticPickupScore(t) {
-  const cat = t.category || "thing";
-  if (!OPPORTUNISTIC_PICKUPS.has(cat)) return null;
-  const distance = typeof t.distance === "number" ? t.distance : 9999;
-  const bearing = typeof t.bearing_deg === "number" ? t.bearing_deg : 999;
-  // Explore mode should not chase side pickups. Only grab things we
-  // are already about to walk through.
-  if (distance > 160 || Math.abs(bearing) > 18) return null;
-  return 20 + distance / 32;
-}
-
-function pickInterestingTarget(state, failedTargets, macro) {
-  const p = state.player;
-  if (!p) return null;
-  for (const [key, expires] of failedTargets) {
-    if (expires <= macro) failedTargets.delete(key);
-  }
-  const candidates = [];
-  for (const t of (state.things_visible || [])) {
-    let pri = thingPriority(t);
-    if (pri === null) pri = opportunisticPickupScore(t);
-    if (pri === null) continue;
-    const cat = t.category || "thing";
-    const distance = typeof t.distance === "number" ? t.distance : 9999;
-    if (distance > 640) continue;
-    const worldBearing = p.angle_deg - t.bearing_deg;
-    const rad = (worldBearing * Math.PI) / 180;
-    const tx = p.x + Math.cos(rad) * distance;
-    const ty = p.y + Math.sin(rad) * distance;
-    const targetKey = "thing:" + cat + ":" + (t.type || "thing") + ":" + Math.round(tx / 64) + "," + Math.round(ty / 64);
-    if (failedTargets.has(targetKey)) continue;
-    candidates.push({
-      kind: cat,
-      type: t.type || "thing",
-      bearing: normAngle(worldBearing),
-      distance,
-      priority: pri,
-      targetKey,
-    });
-  }
-  for (const r of (state.raycasts || [])) {
-    if (!(r.hit in INTEREST_PRIORITY)) continue;
-    if (r.distance > 640) continue;
-    const targetKey = "special:" + r.hit + ":" + Math.round((p.angle_deg - r.bearing_deg) / 15) + ":" + Math.round(r.distance / 64);
-    if (failedTargets.has(targetKey)) continue;
-    candidates.push({
-      kind: r.hit,
-      type: r.thing_type || r.hit,
-      bearing: normAngle(p.angle_deg - r.bearing_deg),
-      distance: r.distance,
-      priority: INTEREST_PRIORITY[r.hit],
-      targetKey,
-    });
-  }
-  if (candidates.length === 0) return null;
-  candidates.sort((a, b) => {
-    if (a.priority !== b.priority) return a.priority - b.priority;
-    return a.distance - b.distance;
-  });
-  return candidates[0];
-}
-
-// Probe a circular patch in the memory canvas and return how
-// "covered" it is (FLOOR/VISITED/...) plus a wallHits count.
-function probePatch(cx, cy, R) {
-  let explored = 0;
-  let total = 0;
-  let wallHits = 0;
-  for (let dy = -R; dy <= R; dy++) {
-    for (let dx = -R; dx <= R; dx++) {
-      if (dx * dx + dy * dy > R * R) continue;
-      const x = Math.round(cx + dx);
-      const y = Math.round(cy + dy);
-      if (x < 0 || y < 0 || x >= SIZE || y >= SIZE) continue;
-      total++;
-      const r = rank[y * SIZE + x];
-      if (r >= RANK.FLOOR) explored++;
-      if (r === RANK.WALL) wallHits++;
-    }
-  }
-  return { explored, total, wallHits, ratio: total > 0 ? explored / total : 1 };
-}
-
-// Pick the best octant to head toward. We score with two concentric
-// probes: a close ring (just past the painted area) where the player
-// has actually observed pixels, and a far ring for "is there room to
-// keep going?". Close ring is weighted more so genuinely-explored
-// directions look "covered" even before the far ring is filled in.
-//
-// \`history\` is a list of the last few { bearing, dist } macro
-// outcomes. Recent picks that produced no movement are excluded from
-// the candidate set entirely so we don't pick them again. If every
-// candidate would be excluded we fall back to a hard turn 90° away
-// from the most recent failed direction.
-function pickLeastExploredOctant(state, history) {
-  const p = state.player;
-  if (!p) return null;
-  const me = worldToPx(p.x, p.y);
-
-  // Two probe rings. Close radius sits at the edge of what raycasts
-  // (max ~1024 units / 128 px) will have painted; far is the
-  // "lookahead" the previous version used.
-  const CLOSE_RADIUS_PX = 20;   // ~160 world units
-  const CLOSE_PROBE_R = 10;
-  const FAR_RADIUS_PX = 56;     // ~448 world units
-  const FAR_PROBE_R = 14;
-
-  // Penalty bookkeeping: any bearing the bot picked in the last 2
-  // macros AND moved < 64 units afterwards is "dead" -- skip it.
-  const dead = new Set();
-  for (const h of history) {
-    if (h && h.dist < 64) dead.add(h.bearing);
-  }
-  const recent = history.length > 0 ? history[history.length - 1] : null;
-
-  const candidates = [];
-  for (let o = 0; o < 8; o++) {
-    const bearing = o * 45;
-    const rad = (bearing * Math.PI) / 180;
-    const close = probePatch(
-      me.px + Math.cos(rad) * CLOSE_RADIUS_PX,
-      me.py - Math.sin(rad) * CLOSE_RADIUS_PX,
-      CLOSE_PROBE_R,
-    );
-    const far = probePatch(
-      me.px + Math.cos(rad) * FAR_RADIUS_PX,
-      me.py - Math.sin(rad) * FAR_RADIUS_PX,
-      FAR_PROBE_R,
-    );
-    // Combined score: close ring weighted 2x. wallHits penalty makes
-    // walls "look covered" so we don't head straight into them.
-    const score =
-      (close.ratio * 2 + far.ratio) / 3 +
-      (close.wallHits + far.wallHits * 0.5) * 0.003 +
-      (recent && recent.dist < STUCK_DIST && Math.abs(angleDelta(bearing, recent.bearing)) < 90 ? 0.8 : 0);
-    candidates.push({
-      bearing,
-      score,
-      closeRatio: close.ratio,
-      farRatio: far.ratio,
-      wallHits: close.wallHits + far.wallHits,
-      isDead: dead.has(bearing),
-    });
-  }
-
-  // First try only "alive" candidates. If none, fall back to the
-  // full set but force a bearing far from any dead one.
-  let alive = candidates.filter((c) => !c.isDead);
-  if (alive.length === 0) alive = candidates;
-  alive.sort((a, b) => {
-    if (a.score !== b.score) return a.score - b.score;
-    // Tiebreak: prefer bearings NOT close to recent failed picks.
-    const recentBearing = history.length > 0 ? history[history.length - 1].bearing : null;
-    if (recentBearing !== null) {
-      const da = Math.abs(angleDelta(a.bearing, recentBearing));
-      const db = Math.abs(angleDelta(b.bearing, recentBearing));
-      if (da !== db) return db - da;
-    }
-    return Math.abs(angleDelta(a.bearing, p.angle_deg)) - Math.abs(angleDelta(b.bearing, p.angle_deg));
-  });
-  const pick = alive[0];
-  pick.debugTop = alive.slice(0, 4).map((c) =>
-    c.bearing + ":" + c.score.toFixed(2) + "/c" + c.closeRatio.toFixed(2) + "/f" + c.farRatio.toFixed(2) + "/w" + c.wallHits + (c.isDead ? "D" : ""),
-  ).join(" ");
-  return pick;
-}
-
-// Extract a CROP x CROP RGBA window centred on the player. Areas
-// outside the source canvas are filled with UNSEEN (opaque black).
-// We also paint a centred arrow showing the player facing so the
-// orientation of the clip is unambiguous.
-function clipAroundPlayer(state) {
-  const p = state.player;
-  if (!p) return null;
-  const me = worldToPx(p.x, p.y);
-  const out = new Uint8Array(CROP * CROP * 4);
-  for (let i = 0; i < CROP * CROP; i++) out[i * 4 + 3] = 255;
-
-  const sx0 = me.px - CROP_HALF;
-  const sy0 = me.py - CROP_HALF;
-  for (let dy = 0; dy < CROP; dy++) {
-    const sy = sy0 + dy;
-    if (sy < 0 || sy >= SIZE) continue;
-    for (let dx = 0; dx < CROP; dx++) {
-      const sx = sx0 + dx;
-      if (sx < 0 || sx >= SIZE) continue;
-      const si = (sy * SIZE + sx) * 4;
-      const oi = (dy * CROP + dx) * 4;
-      out[oi]     = mem[si];
-      out[oi + 1] = mem[si + 1];
-      out[oi + 2] = mem[si + 2];
-      out[oi + 3] = mem[si + 3];
-    }
-  }
-
-  // Player arrow in the dead centre of the crop. A short line in the
-  // facing direction + a red dot on the player pixel.
-  const cx = CROP_HALF;
-  const cy = CROP_HALF;
-  const stamp = (px, py, col) => {
-    if (px < 0 || py < 0 || px >= CROP || py >= CROP) return;
-    const o = (py * CROP + px) * 4;
-    out[o] = col[0]; out[o + 1] = col[1]; out[o + 2] = col[2]; out[o + 3] = 255;
-  };
-  const rad = (p.angle_deg * Math.PI) / 180;
-  const dx = Math.cos(rad);
-  const dy = -Math.sin(rad); // screen y flips
-  for (let i = 1; i <= 8; i++) {
-    stamp(Math.round(cx + dx * i), Math.round(cy + dy * i), COL.PLAYER);
-  }
-  stamp(cx, cy, COL.PLAYER);
-  stamp(cx + 1, cy, COL.PLAYER);
-  stamp(cx - 1, cy, COL.PLAYER);
-  stamp(cx, cy + 1, COL.PLAYER);
-  stamp(cx, cy - 1, COL.PLAYER);
-  return out;
-}
-
-// Closure state for stationary detection. We compare the player's
-// position across consecutive micro ticks to spot "pressing up but
-// not actually moving" wedges that the raycast wall-avoid can't see
-// (fwd ray > 48 because it grazes past a corner).
-let _lastMicroX = null;
-let _lastMicroY = null;
-let _stationaryTicks = 0;
-let _lastMoveIntent = false;
-let _wallAvoidTicks = 0;
-let _avoidBearing = null;
-let _avoidTicks = 0;
-let _avoidStartX = null;
-let _avoidStartY = null;
-let _useCooldownTicks = 0;
-let _postUseForwardTicks = 0;
-const MICRO_MIN_DELTA = 6;   // map units of movement to count as "moving"
-const MICRO_STUCK_TICKS = 3; // back-up after this many stationary ticks
-const AVOID_CLEAR_DIST = 96;
-const WALL_SAFE_DIST = 128;
-const USE_COOLDOWN_TICKS = 8;
-const POST_USE_FORWARD_TICKS = 4;
-
-function microResult(action, moveIntent) {
-  if (!action.startsWith("wall-avoid")) _wallAvoidTicks = 0;
-  _lastMoveIntent = moveIntent;
-  return action;
-}
-
-function isBlockingAhead(ray, distance) {
-  if (!ray || ray.distance >= distance) return false;
-  if (ray.hit === "wall") return true;
-  if (ray.hit !== "thing") return false;
-  const cat = ray.thing_category || "";
-  return cat === "decor" || cat === "barrel" || cat === "enemy";
-}
-
-function bestRay(rays) {
-  let best = rays[0];
-  for (const r of rays) if (!best || r.distance > best.distance) best = r;
-  return best;
-}
-
-function escapeBearing(player, rays, fwd) {
-  const best = bestRay(rays);
-  if (!player || !best) return null;
-  // If the "best" ray is basically the forward ray, this is a corner/pocket.
-  // Pick a hard diagonal escape instead of re-entering the same collision.
-  if (Math.abs(best.bearing_deg) < 15 || (fwd && best.distance < fwd.distance + 96)) {
-    return normAngle(player.angle_deg + 135);
-  }
-  return normAngle(player.angle_deg - best.bearing_deg);
-}
-
-async function microTick(state, target) {
-  if (state.screen !== "playing") {
-    await bot.press("enter");
-    return microResult("menu", false);
-  }
-  const p = state.player;
-  // Position-delta stuck detection. mom is laggy (see get_state docs)
-  // so we use x/y, which are sampled the tic they're read. Reset
-  // on real movement.
-  if (p) {
-    if (_lastMicroX !== null) {
-      const dist = Math.hypot(p.x - _lastMicroX, p.y - _lastMicroY);
-      _stationaryTicks = _lastMoveIntent && dist < MICRO_MIN_DELTA ? _stationaryTicks + 1 : 0;
-    }
-    _lastMicroX = p.x;
-    _lastMicroY = p.y;
-    if (_avoidStartX !== null && Math.hypot(p.x - _avoidStartX, p.y - _avoidStartY) > AVOID_CLEAR_DIST) {
-      _avoidBearing = null;
-      _avoidTicks = 0;
-      _avoidStartX = null;
-      _avoidStartY = null;
-    }
-  }
-
-  const rays = state.raycasts || [];
-  const fwd = rays.find((r) => Math.abs(r.bearing_deg) < 10);
-  if (_useCooldownTicks > 0) _useCooldownTicks--;
-  if (_postUseForwardTicks > 0) {
-    _postUseForwardTicks--;
-    await bot.press("up", 180);
-    return microResult("post-use-fwd", true);
-  }
-
-  const enemies = state.enemies_visible || [];
-  const centred = enemies.find((e) => e.bearing === "center");
-  if (centred) { await bot.press("fire", 100); return microResult("fire", false); }
-  const off = enemies.find((e) => e.bearing === "left" || e.bearing === "far_left") ? "left"
-            : enemies.find((e) => e.bearing === "right" || e.bearing === "far_right") ? "right"
-            : null;
-  if (off) { await bot.press(off, 80); return microResult("face-enemy", false); }
-  const effectiveTarget = _avoidTicks > 0 && _avoidBearing !== null
-    ? { bearing: _avoidBearing, kind: "avoid", distance: 9999 }
-    : target;
-  if (_useCooldownTicks <= 0 && fwd && (fwd.hit === "door" || fwd.hit === "switch" || fwd.hit === "exit") && fwd.distance < 160 && Math.abs(fwd.bearing_deg) < 8) {
-    await bot.press("use", 30);
-    await bot.press("up", 120);
-    _useCooldownTicks = USE_COOLDOWN_TICKS;
-    _postUseForwardTicks = POST_USE_FORWARD_TICKS;
-    return microResult("use", true);
-  }
-
-  // Pinned: the engine isn't letting us move regardless of which key
-  // we press. Back up to break contact with whatever geometry has
-  // us stuck (corner, doorframe, decoration). Reset the counter
-  // afterwards so we get one clean tick to re-evaluate.
-  if (_stationaryTicks >= MICRO_STUCK_TICKS) {
-    const best = bestRay(rays);
-    const escape = escapeBearing(p, rays, fwd);
-    if (p && escape !== null) {
-      _avoidBearing = escape;
-      _avoidTicks = 8;
-      _avoidStartX = p.x;
-      _avoidStartY = p.y;
-    }
-    await bot.press("down", 160);
-    if (p && escape !== null) {
-      await bot.press(angleDelta(escape, p.angle_deg) > 0 ? "left" : "right", 220);
-    } else {
-      await bot.press("right", 220);
-    }
-    _stationaryTicks = 0;
-    return microResult("unwedge:" + (best ? best.bearing_deg.toFixed(0) : "?"), true);
-  }
-
-  if (isBlockingAhead(fwd, WALL_SAFE_DIST)) {
-    _wallAvoidTicks++;
-    const best = bestRay(rays);
-    const escape = escapeBearing(p, rays, fwd);
-    if (p && escape !== null) {
-      _avoidBearing = escape;
-      _avoidTicks = 8;
-      if (_avoidStartX === null) {
-        _avoidStartX = p.x;
-        _avoidStartY = p.y;
-      }
-    }
-    if (_wallAvoidTicks >= 4) {
-      await bot.press("down", 220);
-      if (p && escape !== null) {
-        await bot.press(angleDelta(escape, p.angle_deg) > 0 ? "left" : "right", 260);
-      } else {
-        await bot.press("right", 260);
-      }
-      return microResult("wall-escape:" + (best ? best.bearing_deg.toFixed(0) : "?"), true);
-    }
-    if (p && escape !== null) {
-      await bot.press(angleDelta(escape, p.angle_deg) > 0 ? "left" : "right", 220);
-    } else {
-      await bot.press(best.bearing_deg < 0 ? "left" : "right", Math.min(260, 120 + Math.abs(best.bearing_deg) * 4));
-    }
-    return microResult("wall-avoid:" + (best ? best.bearing_deg.toFixed(0) : "?"), false);
-  }
-  if (!isBlockingAhead(fwd, WALL_SAFE_DIST) && _avoidTicks > 0) _avoidTicks--;
-  if (_avoidTicks <= 0 && _avoidStartX === null) _avoidBearing = null;
-  if (effectiveTarget && state.player) {
-    const err = angleDelta(effectiveTarget.bearing, state.player.angle_deg);
-    if (_useCooldownTicks <= 0 && (effectiveTarget.kind === "door" || effectiveTarget.kind === "switch" || effectiveTarget.kind === "exit") && effectiveTarget.distance < 180 && Math.abs(err) < 25) {
-      await bot.press("use", 40);
-      await bot.press("up", 120);
-      _useCooldownTicks = USE_COOLDOWN_TICKS;
-      _postUseForwardTicks = POST_USE_FORWARD_TICKS;
-      return microResult("use-target", true);
-    }
-    if (Math.abs(err) > TURN_TOLERANCE_DEG) {
-      const turnKey = err > 0 ? "left" : "right";
-      const hold = Math.min(240, Math.max(70, Math.abs(err) * 2.6));
-      await bot.press(turnKey, hold);
-      if (Math.abs(err) < 60 && !isBlockingAhead(fwd, WALL_SAFE_DIST)) {
-        await bot.press("up", 90);
-        return microResult("steer-fwd:" + effectiveTarget.kind + ":" + turnKey + ":" + err.toFixed(0) + ":" + hold.toFixed(0), true);
-      }
-      return microResult("steer:" + effectiveTarget.kind + ":" + turnKey + ":" + err.toFixed(0) + ":" + hold.toFixed(0), false);
-    }
-  }
-  if (isBlockingAhead(fwd, WALL_SAFE_DIST)) {
-    await bot.press("right", 180);
-    return microResult("blocked-turn", false);
-  }
-  await bot.press("up", 130);
-  return microResult("fwd", true);
-}
-
-// Per-macro history: { bearing, dist } for the last few macros. The
-// picker uses this to exclude bearings that just produced no
-// movement, breaking the "same direction every macro" loop.
-const history = [];
-const HISTORY_LEN = 3;
-const failedTargets = new Map();
-// Distance below which a macro is considered "stuck".
-const STUCK_DIST = 64;
-
-for (let macro = 0; macro < STEPS; macro++) {
-  let s = await bot.getState();
-  if (s.screen !== "playing" && s.screen !== "automap") {
-    await bot.press("enter");
-    await bot.sleep(200);
-    continue;
-  }
-  recordObservation(s, macro, 0);
-  const interesting = pickInterestingTarget(s, failedTargets, macro);
-  let pick = interesting || pickLeastExploredOctant(s, history);
-
-  // Hard-stuck escape: if the last 2 macros barely moved, force a
-  // bearing >= 90° off the most recent pick regardless of score.
-  const stuckRun = history.slice(-2).filter((h) => h.dist < STUCK_DIST).length;
-  if (stuckRun >= 2 && history.length > 0) {
-    const last = history[history.length - 1].bearing;
-    const forced = (last + 135) % 360; // sharp turn, not a U-turn
-    pick = { bearing: forced, score: -1, closeRatio: 0, farRatio: 0, wallHits: 0, forced: true };
-  }
-  const target = pick ? { bearing: pick.bearing, kind: pick.kind || (pick.forced ? "forced" : "explore"), distance: pick.distance || 9999 } : null;
-  const targetBearing = target ? target.bearing : null;
-
-  const startPose = { x: s.player.x, y: s.player.y };
-  const actionCounts = {};
-  await bot.log(
-    "macro " + macro +
-    " pose=(" + s.player.x.toFixed(0) + "," + s.player.y.toFixed(0) + ")@" +
-    s.player.angle_deg.toFixed(0) + "° things_tracked=" + things.size +
-    " stuckRun=" + stuckRun + " " +
-    (pick
-      ? "-> head " + targetBearing + "°" +
-        (interesting ? " [TARGET " + interesting.kind + ":" + interesting.type + " d=" + interesting.distance.toFixed(0) + "]" :
-        (pick.forced ? " [FORCED 135° off last pick]" :
-          " (close=" + pick.closeRatio.toFixed(2) +
-          ", far=" + pick.farRatio.toFixed(2) +
-          ", wallHits=" + pick.wallHits + ")"))
-      : "no target"),
-  );
-  if (pick && pick.debugTop) await bot.log("  candidates " + pick.debugTop);
-  if (target) {
-    const me = worldToPx(s.player.x, s.player.y);
-    await bot.log(
-      "  target kind=" + target.kind + " bearing=" + target.bearing.toFixed(0) +
-      " dist=" + target.distance.toFixed(0) + " playerPx=(" + me.px + "," + me.py + ")",
-    );
-  }
-
-  for (let t = 0; t < TICKS_PER_MACRO; t++) {
-    s = await bot.getState();
-    if (s.screen === "dead" || s.screen === "finale") {
-      return "ended on " + s.screen + " after " + macro + " macros";
-    }
-    recordObservation(s, macro, t);
-    const before = s.player ? { x: s.player.x, y: s.player.y, angle: s.player.angle_deg } : null;
-    const rays = s.raycasts || [];
-    const fwd = rays.find((r) => Math.abs(r.bearing_deg) < 10);
-    let best = rays[0];
-    for (const r of rays) if (!best || r.distance > best.distance) best = r;
-    const errBefore = before && target ? angleDelta(target.bearing, before.angle) : null;
-    const action = await microTick(s, target);
-    actionCounts[action] = (actionCounts[action] || 0) + 1;
-    if (DEBUG_TICKS && before) {
-      const me = worldToPx(before.x, before.y);
-      await bot.log(
-        "    t" + t +
-        " pose=(" + before.x.toFixed(0) + "," + before.y.toFixed(0) + ")@" + before.angle.toFixed(0) +
-        " px=(" + me.px + "," + me.py + ")" +
-        " target=" + (target ? target.bearing.toFixed(0) + "/" + target.kind : "none") +
-        " avoid=" + (_avoidTicks > 0 && _avoidBearing !== null ? _avoidBearing.toFixed(0) + ":" + _avoidTicks : "none") +
-        " useCd=" + _useCooldownTicks + "/post=" + _postUseForwardTicks +
-        " err=" + (errBefore === null ? "n/a" : errBefore.toFixed(0)) +
-        " fwd=" + (fwd ? fwd.hit + ":" + fwd.distance.toFixed(0) + ":b" + fwd.bearing_deg.toFixed(0) + (fwd.thing_category ? ":" + fwd.thing_category : "") : "?") +
-        " best=" + (best ? best.hit + ":" + best.distance.toFixed(0) + ":b" + best.bearing_deg.toFixed(0) : "?") +
-        " block96=" + (isBlockingAhead(fwd, 96) ? "Y" : "n") +
-        " action=" + action,
-      );
-    }
-    await bot.sleep(TICK_PAUSE_MS);
-  }
-
-  // Record this macro's outcome for future history-based decisions.
-  const movedDist = Math.hypot(s.player.x - startPose.x, s.player.y - startPose.y);
-  history.push({ bearing: targetBearing, dist: movedDist });
-  if (history.length > HISTORY_LEN) history.shift();
-  if (pick && pick.targetKey && movedDist < 80) {
-    failedTargets.set(pick.targetKey, macro + 3);
-  }
-  await bot.log(
-    "  macro " + macro + " moved " + movedDist.toFixed(0) +
-    " units actions=" + JSON.stringify(actionCounts),
-  );
-
-  if (macro % LOG_EVERY === 0 || macro === STEPS - 1) {
-    const clip = clipAroundPlayer(s);
-    if (clip) {
-      const png = await bot.encodePng(CROP, CROP, clip);
-      await bot.logImage(
-        png,
-        "macro " + macro + " clip " + CROP + "x" + CROP + "px (" +
-        (CROP * SCALE) + "x" + (CROP * SCALE) + " units) — things=" + things.size,
-      );
-    }
-  }
-}
-
-return "finished " + STEPS + " macros";
-
-`;
-
-const INSPECT_BOT = `// Dump a single state snapshot and the current frame, then quit.
-// Useful for sanity-checking what the engine exposes.
-await bot.log(JSON.stringify(await bot.getState(), null, 2));
-// The screenshot lands in the collapsible image panel on the right.
-const shot = await bot.screenshot();
-await bot.logImage(shot, "inspect: current frame");
-`;
-
-// Vision-LLM bot: opens the automap, screenshots it, asks Workers AI
-// where to go, and moves in that direction. Requires the optional
-// `ai` namespace (host worker must have the AI binding configured).
-//
-// API used:
-//   await bot.screenshot()  -> { data: base64-png, mimeType }
-//   await ai.run(model, input)
-const AI_NAV_BOT = `// Hybrid AI navigation with closed-loop steering.
-//
-// One macro consult: open the automap, screenshot it, ask the vision
-// LLM for a *player-relative* turn (AHEAD / SOFT_LEFT / HARD_LEFT /
-// SOFT_RIGHT / HARD_RIGHT / BACK). Combine it with the player's current
-// facing to compute a target world bearing.
-//
-// Many micro ticks: read get_state every tick, compute the angular
-// error between the current player.angle_deg and the target bearing,
-// and correct it (turn left/right) before walking forward. Combat /
-// door / pinned / stuck / wall overrides preempt steering.
-//
-// Idea: the LLM is slow and expensive, so we only use it to answer the
-// hard question — "given the whole map, which way should I be heading?"
-// Everything else (shooting enemies in the FOV, opening doors, not
-// walking into walls) is done locally from the raycast + thing data
-// in get_state, tick-by-tick.
-//
-// Requires the host worker to have the optional Workers AI binding
-// configured (see doom-player wrangler.jsonc \`ai\` block). Without
-// it, \`ai\` is undefined in this sandbox and the call below throws.
-
-const MODEL = "@cf/meta/llama-3.2-11b-vision-instruct";
-const STEPS = 5;                // macro consults of the vision LLM
-const TICKS_PER_MACRO = 24;     // micro ticks between consults
-const TICK_MS = 200;
-const TURN_TOLERANCE_DEG = 18;  // dead-band: don't bother correcting <18°
-const TURN_HOLD_MIN_MS = 140;   // shortest turn tap
-const TURN_HOLD_MAX_MS = 500;   // longest single turn tap (big errors)
-// Every field returned by get_state is a snapshot of the last completed
-// 35Hz engine tic (see the \`get_state\` tool description in
-// src/app/lib/webmcp.tsx). After a press_key our keydown -> ticcmd ->
-// thrust pipeline can leave the next get_state still showing the
-// previous tic's pose, raycasts, momx/momy, things, etc.
-//
-// momx/momy are the most visibly laggy because they require the thrust
-// step to have run, but x/y/angle/raycasts can also briefly trail.
-// Position eventually becomes ground truth because the displacement
-// from a press accumulates over multiple tics, so deltas between two
-// consecutive reads are a reliable "did anything happen?" signal even
-// when one read is stale. We use position-delta-over-multiple-ticks
-// for wedge detection; momentum is just logged for context.
-const STUCK_POS_EPS = 4;        // units of pos delta below this = stuck
-const STUCK_TICKS = 3;          // consecutive stuck ticks before unsticking
-// Kept for the per-tick log and steady-state speed estimates only;
-// not used for wedge detection any more.
-const STUCK_MOM_EPS = 0.5;
-
-// @cf/meta/llama-3.2-11b-vision-instruct requires that you agree with their terms
-await ai.run(MODEL, { prompt: "agree" }).catch(() => {});
-
-// ── Heading helpers ──────────────────────────────────────────────────
-//
-// Earlier versions of this bot treated the automap as world-axis-
-// aligned and asked the LLM for a screen-space 3x3 cell, then converted
-// that to an absolute world bearing. That was wrong in practice: the
-// player ARROW on the automap rotates with the player's facing, so the
-// LLM naturally reads the map relative to the arrow ("the open corridor
-// is ahead and to the right of the player"). Converting its answer as
-// an absolute world direction made the bot turn the wrong way whenever
-// the player wasn't already facing north.
-//
-// We now ask the LLM for a *player-relative* turn (AHEAD / SOFT_LEFT /
-// HARD_LEFT / SOFT_RIGHT / HARD_RIGHT / BACK) and compute the target
-// world bearing as \`player.angle + relative_offset\`. No screen-to-world
-// conversion needed.
-const RELATIVE_TURNS = {
-  AHEAD:       0,
-  SOFT_LEFT:  45,
-  HARD_LEFT:  90,
-  BACK:      180,
-  HARD_RIGHT: -90,
-  SOFT_RIGHT: -45,
-};
-
-// Signed angular delta in degrees, result in (-180, 180].
-function angleDelta(targetDeg, currentDeg) {
-  let d = (targetDeg - currentDeg) % 360;
-  if (d > 180) d -= 360;
-  if (d <= -180) d += 360;
-  return d;
-}
-
-function normalizeAngle(deg) {
-  return ((deg % 360) + 360) % 360;
-}
-
-// Cardinal label for a world bearing (Doom convention: 0°=E, 90°=N).
-// Lets us log human-readable directions alongside raw degrees so traces
-// stay readable without doing degree arithmetic in your head.
-function compassLabel(deg) {
-  const d = normalizeAngle(deg);
-  const labels = [
-    ["E", 0], ["NE", 45], ["N", 90], ["NW", 135],
-    ["W", 180], ["SW", 225], ["S", 270], ["SE", 315],
-  ];
-  let best = labels[0];
-  let bestDelta = 360;
-  for (const [name, ref] of labels) {
-    const delta = Math.min(
-      Math.abs(d - ref),
-      360 - Math.abs(d - ref),
-    );
-    if (delta < bestDelta) {
-      bestDelta = delta;
-      best = [name, ref];
-    }
-  }
-  return best[0];
-}
-
-// Linearly interpolate the engine's 8 forward raycasts to estimate
-// what's at an arbitrary player-relative bearing. Used by the veto
-// check after the LLM picks a turn — if the predicted slot is a close
-// wall we override to BACK rather than walking into geometry.
-function predictRayAt(rays, relBearingDeg) {
-  if (!rays || rays.length === 0) return null;
-  // Find the two nearest rays by bearing.
-  let nearest = rays[0];
-  let nearestDelta = Infinity;
-  for (const r of rays) {
-    const delta = Math.abs(r.bearing_deg - relBearingDeg);
-    if (delta < nearestDelta) {
-      nearestDelta = delta;
-      nearest = r;
-    }
-  }
-  return nearest;
-}
-
-// Build a compact, structured prose digest of the current engine
-// state for the LLM. All directional fields are *player-relative*
-// (matches the turn-direction question we're about to ask). The
-// digest is intentionally line-oriented so the model can scan it.
-function buildStatePrompt(state, history) {
-  const p = state.player;
-  const facingStr = p
-    ? \`\${p.angle_deg.toFixed(0)}° (\${compassLabel(p.angle_deg)})\`
-    : "(unknown)";
-  const poseStr = p
-    ? \`(\${p.x.toFixed(0)}, \${p.y.toFixed(0)})\`
-    : "(unknown)";
-
-  const rays = state.raycasts || [];
-  const raysSorted = [...rays].sort((a, b) => a.bearing_deg - b.bearing_deg);
-  const rayLines = raysSorted.map((r) => {
-    const sign = r.bearing_deg >= 0 ? "+" : "";
-    const extra = r.thing_type ? \` (\${r.thing_type})\` : "";
-    return \`  \${sign}\${r.bearing_deg.toFixed(0).padStart(3, " ")}°: \${r.hit} @ \${r.distance.toFixed(0)}\${extra}\`;
-  }).join("\\n");
-
-  const things = (state.things_visible || []).slice(0, 8).map((t) => {
-    const sign = t.bearing_deg >= 0 ? "+" : "";
-    return \`\${t.type} @ \${sign}\${t.bearing_deg.toFixed(0)}°/\${t.distance.toFixed(0)}\`;
-  }).join(", ");
-
-  const enemies = (state.enemies_visible || []).map(
-    (e) => \`\${e.type} @ \${e.bearing}/\${e.distance}\`,
-  ).join(", ");
-
-  const hud = state.hud;
-  const hudStr = \`hp=\${hud.health} armor=\${hud.armor} ammo=\${hud.ammo}(\${hud.ammo_type}) weapon=\${hud.weapon} keys=[\${(hud.keys || []).join(",")}]\`;
-
-  const historyLines = [];
-  if (history && history.recent && history.recent.length > 0) {
-    for (const h of history.recent) {
-      historyLines.push(
-        \`  macro \${h.macro}: picked \${h.turn}, moved \${h.dist.toFixed(0)} units\${h.pinned ? " (pinned)" : ""}\`,
-      );
-    }
-  }
-
-  return [
-    "ENGINE STATE (all bearings are relative to player facing; -=left, +=right):",
-    \`  facing: \${facingStr}\`,
-    \`  pose:   \${poseStr}\`,
-    \`  hud:    \${hudStr}\`,
-    "  forward-cone raycasts:",
-    rayLines || "    (none)",
-    \`  things visible: \${things || "(none)"}\`,
-    \`  enemies visible: \${enemies || "(none)"}\`,
-    historyLines.length > 0
-      ? "RECENT MACRO HISTORY:\\n" + historyLines.join("\\n")
-      : "RECENT MACRO HISTORY: (this is the first macro)",
-  ].join("\\n");
-}
-
-// Convert base64 PNG -> number[] (Workers AI vision input shape).
-function base64ToBytes(b64) {
-  const bin = atob(b64);
-  const out = new Array(bin.length);
-  for (let i = 0; i < bin.length; i++) out[i] = bin.charCodeAt(i);
-  return out;
-}
-
-// Capture the automap as a screenshot. The automap is a toggle (Tab),
-// so we open it, wait a tic for redraw, snap, then close it again so
-// regular movement keys go back to controlling the player.
-async function snapAutomap(currentScreen) {
-  if (currentScreen !== "automap") {
-    await bot.press("tab");
-    await bot.sleep(150);
-  }
-  const shot = await bot.screenshot();
-  await bot.press("tab");
-  await bot.sleep(100);
-  return shot;
-}
-
-// Ask the vision LLM which way to turn next, relative to the player
-// arrow on the automap. The prompt fuses the rendered automap (image)
-// with a structured digest of the engine state (text), so the model
-// can cross-reference what it "sees" with what raycasts actually show.
-async function askMacroTurn(state, shot, history) {
-  const stateBlock = buildStatePrompt(state, history);
-  const instructions =
-    "You are guiding a DOOM bot. The IMAGE is the in-game automap: " +
-    "white lines are explored walls; the small white triangle at the " +
-    "centre is the PLAYER and its tip points the way the player faces. " +
-    "Black space adjacent to white walls is unexplored territory.\\n\\n" +
-    stateBlock +
-    "\\n\\nDecide where the player should head next, *relative to the " +
-    "arrow's current facing*. Use BOTH the map (for big-picture " +
-    "exploration) AND the raycasts (for what's physically reachable " +
-    "this second). Prefer a direction where the raycasts are open " +
-    "(distance > 100, hit != wall). Avoid picking a direction the " +
-    "raycasts show as a close wall. If the recent history shows the " +
-    "bot was pinned moving the same way, pick a DIFFERENT direction " +
-    "this time.\\n\\n" +
-    "Reply with EXACTLY one token from this set, nothing else: " +
-    "AHEAD, SOFT_LEFT, HARD_LEFT, SOFT_RIGHT, HARD_RIGHT, BACK.\\n" +
-    "  AHEAD       = keep current facing\\n" +
-    "  SOFT_LEFT   = rotate ~45° counter-clockwise\\n" +
-    "  HARD_LEFT   = rotate ~90° counter-clockwise\\n" +
-    "  SOFT_RIGHT  = rotate ~45° clockwise\\n" +
-    "  HARD_RIGHT  = rotate ~90° clockwise\\n" +
-    "  BACK        = turn around (~180°)";
-
-  const t0 = Date.now();
-  const resp = await ai.run(MODEL, {
-    image: base64ToBytes(shot.data),
-    prompt: instructions,
-    max_tokens: 12,
-  });
-  const elapsedMs = Date.now() - t0;
-  const text = (resp && typeof resp === "object" && typeof resp.response === "string")
-    ? resp.response.trim()
-    : String(resp);
-  const upper = text.toUpperCase();
-  // Longest tokens first so "HARD_LEFT" doesn't match the "LEFT" branch.
-  let pick = "AHEAD";
-  for (const k of ["HARD_LEFT", "HARD_RIGHT", "SOFT_LEFT", "SOFT_RIGHT", "BACK", "AHEAD"]) {
-    if (upper.includes(k)) { pick = k; break; }
-  }
-  // Log the full prompt + raw response so the trace alone is enough
-  // to reconstruct what the LLM saw and replied.
-  await bot.log(\`  ai prompt (\${instructions.length} chars):\`);
-  for (const line of instructions.split("\\n")) await bot.log(\`    | \${line}\`);
-  await bot.log(\`  ai (\${elapsedMs}ms) raw=\${JSON.stringify(text)} -> \${pick}\`);
-  return { pick, elapsedMs, rawText: text, prompt: instructions };
-}
-
-// Safety net: if the LLM's chosen direction lands on a close wall
-// according to the engine's raycasts, override it. The veto returns
-// either the original pick (no change) or a replacement that points
-// at the deepest open ray; logging shows which case fired.
-function vetoTurn(pick, state) {
-  const offset = RELATIVE_TURNS[pick] ?? 0;
-  const rays = state.raycasts || [];
-  if (rays.length === 0) return { pick, vetoed: false };
-  // Forward-cone rays only; BACK is never vetoed (we trust the LLM
-  // on "turn around" because the engine's forward rays say nothing
-  // about what's behind the player).
-  if (pick === "BACK") return { pick, vetoed: false };
-
-  const sample = predictRayAt(rays, offset);
-  if (!sample) return { pick, vetoed: false };
-  // The "blocked" threshold is generous on purpose — we only veto
-  // when the engine is very confident the chosen lane is unwalkable.
-  if (sample.hit !== "wall" || sample.distance >= 40) {
-    return { pick, vetoed: false, sample };
-  }
-
-  // Pick the deepest open ray (any non-wall, or wall with > 100
-  // distance) and translate its bearing back into a label.
-  let best = rays[0];
-  for (const r of rays) if (r.distance > best.distance) best = r;
-  let replacement = "AHEAD";
-  const b = best.bearing_deg;
-  if (b >= 67) replacement = "HARD_LEFT";
-  else if (b >= 22) replacement = "SOFT_LEFT";
-  else if (b <= -67) replacement = "HARD_RIGHT";
-  else if (b <= -22) replacement = "SOFT_RIGHT";
-  // If the deepest open ray is *also* close, fall back to BACK.
-  if (best.hit === "wall" && best.distance < 40) replacement = "BACK";
-  return {
-    pick: replacement,
-    vetoed: true,
-    sample,
-    reason: \`ray at \${offset}° is \${sample.hit}@\${sample.distance.toFixed(0)}; deepest ray = \${b.toFixed(0)}°@\${best.distance.toFixed(0)} (\${best.hit})\`,
-  };
-}
-
-// Convert a player-relative turn label into an absolute target world
-// bearing using the player's current facing.
-function turnToTargetBearing(turn, player) {
-  if (!player) return null;
-  const offset = RELATIVE_TURNS[turn] ?? 0;
-  // \`offset\` is in Doom's CCW-positive convention (LEFT = +45°).
-  return normalizeAngle(player.angle_deg + offset);
-}
-
-// Scale turn-tap duration with the absolute heading error so big
-// macro errors don't take 20 ticks to close.
-function turnHoldFor(errDeg) {
-  const abs = Math.min(180, Math.abs(errDeg));
-  // Linear: TURN_HOLD_MIN_MS at 18° (tolerance), TURN_HOLD_MAX_MS at 90°+.
-  const t = Math.min(1, Math.max(0, (abs - TURN_TOLERANCE_DEG) / (90 - TURN_TOLERANCE_DEG)));
-  return Math.round(TURN_HOLD_MIN_MS + t * (TURN_HOLD_MAX_MS - TURN_HOLD_MIN_MS));
-}
-
-// Doom's "use" is a toggle on doors — every press flips the door's
-// open/close state. Hammering use@30ms every tick means we keep
-// closing the door we just opened. Block consecutive use presses
-// for this many ticks so the engine has time to animate the door.
-const USE_COOLDOWN_TICKS = 8;
-let _useCooldown = 0;
-
-// Deterministic one-tick policy. Combat / doors preempt steering.
-// Wall avoidance is biased toward the macro target side. The caller
-// hands us a stuck counter (derived from position delta — see the
-// STUCK_POS_EPS note above) and an \`actuallyMoving\` flag so we can
-// distinguish "engine says mom=0 but we just teleported 60 units" from
-// "engine says mom=0 and we genuinely haven't moved".
-// Returns a short string describing which branch fired, so the caller
-// can log it for offline analysis ("why did the bot do X on tick Y?").
-async function microTick(state, targetBearing, stuckTicks, actuallyMoving) {
-  if (_useCooldown > 0) _useCooldown -= 1;
-  if (state.screen !== "playing") {
-    await bot.press("enter");
-    return \`menu(enter) screen=\${state.screen}\`;
-  }
-
-  // 1. Combat: shoot centred enemies, turn toward off-centre ones.
-  const enemies = state.enemies_visible || [];
-  const centred = enemies.find((e) => e.bearing === "center");
-  if (centred) {
-    await bot.press("fire", 200);
-    return \`fire @\${centred.type}\`;
-  }
-  const turnTowardEnemy = enemies.find((e) => e.bearing === "left" || e.bearing === "far_left")
-    ? "left"
-    : enemies.find((e) => e.bearing === "right" || e.bearing === "far_right")
-    ? "right"
-    : null;
-  if (turnTowardEnemy) {
-    await bot.press(turnTowardEnemy, 120);
-    return \`face-enemy \${turnTowardEnemy}\`;
-  }
-
-  // 2. Doors / switches close ahead -> activate and step through.
-  //    Two guards before pressing \`use\`:
-  //      a. the door must be near-centred in the FOV (otherwise we're
-  //         not actually facing it; let steering align us first).
-  //      b. respect a cooldown — \`use\` toggles the door, so spamming
-  //         it every tick keeps re-closing what we just opened.
-  const rays = state.raycasts || [];
-  const fwd = rays.find((r) => Math.abs(r.bearing_deg) < 10);
-  if (
-    fwd &&
-    (fwd.hit === "door" || fwd.hit === "switch") &&
-    fwd.distance < 80 &&
-    Math.abs(fwd.bearing_deg) < 8 &&
-    _useCooldown === 0
-  ) {
-    // Single short tap; door animation runs even while we walk
-    // forward, so don't burn ticks holding the key.
-    await bot.press("use", 30);
-    await bot.press("up", 200);
-    _useCooldown = USE_COOLDOWN_TICKS;
-    return \`use \${fwd.hit}@\${fwd.distance.toFixed(0)} (cooldown=\${USE_COOLDOWN_TICKS})\`;
-  }
-
-  // 3. Pinned: every nearby forward ray is a close wall. Pure rotation
-  //    won't help — the player needs to physically retreat first. This
-  //    catches the "wedged in a corner" case where the unstick turn
-  //    below would just spin in place.
-  const fwdRays = (rays || []).filter((r) => Math.abs(r.bearing_deg) < 45);
-  const pinned =
-    fwdRays.length > 0 &&
-    fwdRays.every((r) => r.hit === "wall" && r.distance < 32);
-  if (pinned) {
-    await bot.press("down", 350);
-    return \`pinned: back up (rays=\${fwdRays.map((r) => r.distance.toFixed(0)).join(",")})\`;
-  }
-
-  // 4. Stuck (no momentum for several ticks) -> escape: back up
-  //    *then* turn. Backing up reliably breaks contact with whatever
-  //    geometry we wedged into; the turn happens on the next tick.
-  if (stuckTicks >= STUCK_TICKS) {
-    await bot.press("down", 250);
-    await bot.press("right", 250);
-    return \`unstick (stuck=\${stuckTicks})\`;
-  }
-
-  // 5. Wall in our face -> turn toward the deepest open ray, but
-  //    *prefer* the side closer to the macro target when both sides
-  //    are roughly equal.
-  if (fwd && fwd.hit === "wall" && fwd.distance < 48) {
-    const targetErr =
-      targetBearing !== null && state.player
-        ? angleDelta(targetBearing, state.player.angle_deg)
-        : 0;
-    let best = rays[0];
-    for (const r of rays) if (r.distance > best.distance) best = r;
-    // If the macro target is more than 45° off, override the
-    // deepest-ray pick with the target side; it's better to grind a
-    // tic and turn correctly than walk away from where we want to go.
-    const targetBiased = Math.abs(targetErr) > 45;
-    const dir = targetBiased
-      ? targetErr > 0
-        ? "left"
-        : "right"
-      : best.bearing_deg < 0
-      ? "left"
-      : "right";
-    await bot.press(dir, 220);
-    return \`wall@\${fwd.distance.toFixed(0)} turn \${dir} (\${targetBiased ? "target-bias" : \`deep-ray@\${best.bearing_deg.toFixed(0)}\`})\`;
-  }
-
-  // 6. Closed-loop steering toward the macro target bearing.
-  if (targetBearing !== null && state.player) {
-    const err = angleDelta(targetBearing, state.player.angle_deg);
-    if (Math.abs(err) > TURN_TOLERANCE_DEG) {
-      // Doom's angles: +y is 90°, -y is 270°. A positive \`err\` means
-      // we need to rotate counter-clockwise, which is the \`left\` key.
-      const hold = turnHoldFor(err);
-      const dir = err > 0 ? "left" : "right";
-      await bot.press(dir, hold);
-      return \`steer \${dir} \${hold}ms (err=\${err.toFixed(0)}°)\`;
-    }
-  }
-
-  // 7. Heading is good (or no target): walk forward — unless we're
-  //    *already* stationary with a wall in range. The default wall-
-  //    avoid branch above only fires at distance<48, but a player
-  //    facing wall@80 with no actual position progress will sit there
-  //    pressing "up" against geometry forever. Escalate to a sideways
-  //    turn. We use \`actuallyMoving\` (position-delta based) here
-  //    rather than mom, because mom lags behind the engine.
-  const fwdRay = fwd; // alias for clarity
-  if (
-    !actuallyMoving &&
-    stuckTicks >= 1 &&
-    fwdRay &&
-    fwdRay.hit === "wall" &&
-    fwdRay.distance < 120
-  ) {
-    // Pick the deepest open ray and turn that way; same as branch 5
-    // but triggered earlier because spd=0 + close-ish wall is a
-    // strong "the engine won't let me through" signal.
-    let best = rays[0];
-    for (const r of rays) if (r.distance > best.distance) best = r;
-    const dir = best.bearing_deg < 0 ? "left" : "right";
-    await bot.press(dir, 220);
-    return \`stalled@wall\${fwdRay.distance.toFixed(0)} turn \${dir} (deep-ray@\${best.bearing_deg.toFixed(0)}@\${best.distance.toFixed(0)})\`;
-  }
-
-  await bot.press("up", 250);
-  return targetBearing === null ? "fwd (no target)" : "fwd (on-bearing)";
-}
-
-// History of recent macro outcomes; fed back to the LLM so it can
-// recognise it was pinned and pick a different direction next time.
-// Keep the last 3 entries to avoid bloating the prompt.
-const history = { recent: [] };
-
-for (let macro = 0; macro < STEPS; macro++) {
-  let s = await bot.getState();
-  await bot.log("macro", macro, "screen:", s.screen, "hp:", s.hud.health,
-    "pose:", s.player ? \`(\${s.player.x.toFixed(0)},\${s.player.y.toFixed(0)})@\${s.player.angle_deg.toFixed(0)}° (\${compassLabel(s.player.angle_deg)})\` : "?");
-
-  if (s.screen !== "playing" && s.screen !== "automap") {
-    await bot.press("enter");
-    await bot.sleep(200);
-    continue;
-  }
-
-  // --- ONE LLM consult per macro step: image + state digest. ---
-  const shot = await snapAutomap(s.screen);
-  // Surface the automap to the UI's debug-image panel so a human
-  // watching the run can see exactly what the LLM saw.
-  await bot.logImage(shot, \`macro \${macro} automap (pose \${s.player ? \`(\${s.player.x.toFixed(0)},\${s.player.y.toFixed(0)})@\${s.player.angle_deg.toFixed(0)}°\` : "?"})\`);
-  const llmResult = await askMacroTurn(s, shot, history);
-  let turn = llmResult.pick;
-
-  // Safety net: if the LLM picked a direction the raycasts say is a
-  // close wall, override. Logs both versions so we can tell whether
-  // the veto fires too aggressively.
-  const veto = vetoTurn(turn, s);
-  if (veto.vetoed) {
-    await bot.log(\`  veto: \${llmResult.pick} -> \${veto.pick} (\${veto.reason})\`);
-    turn = veto.pick;
-  } else if (veto.sample) {
-    await bot.log(
-      \`  veto: kept \${llmResult.pick} (ray@\${(RELATIVE_TURNS[llmResult.pick] ?? 0)}° = \${veto.sample.hit}@\${veto.sample.distance.toFixed(0)})\`,
-    );
-  }
-
-  const targetBearing = turnToTargetBearing(turn, s.player);
-  await bot.log(
-    "macro turn:", turn,
-    targetBearing === null
-      ? "(no player pose)"
-      : \`-> target bearing \${targetBearing.toFixed(0)}° (\${compassLabel(targetBearing)}) from facing \${s.player ? s.player.angle_deg.toFixed(0) : "?"}° (\${s.player ? compassLabel(s.player.angle_deg) : "?"})\`,
-  );
-
-  // --- Closed-loop micro ticks: check state, correct heading. ---
-  let stuckTicks = 0;
-  const startPose = s.player ? { x: s.player.x, y: s.player.y } : null;
-  // Last observed pose, used to compute position delta tick-to-tick.
-  // Position is the ground-truth movement signal (mom is laggy).
-  let lastPos = s.player ? { x: s.player.x, y: s.player.y } : null;
-  const branchCounts = {};
-  for (let t = 0; t < TICKS_PER_MACRO; t++) {
-    s = await bot.getState();
-    if (s.screen === "dead" || s.screen === "finale") {
-      await bot.log("ending screen reached:", s.screen);
-      return \`ended on \${s.screen} after \${macro} macro steps\`;
-    }
-    // Position delta since last get_state. This is what we actually
-    // trust for "are we moving?" — mom lags 1-2 engine tics behind
-    // press_key, but x/y are sampled the same tic they're read.
-    const posDelta =
-      s.player && lastPos
-        ? Math.hypot(s.player.x - lastPos.x, s.player.y - lastPos.y)
-        : 0;
-    const actuallyMoving = posDelta >= STUCK_POS_EPS;
-    stuckTicks = actuallyMoving ? 0 : stuckTicks + 1;
-    if (s.player) lastPos = { x: s.player.x, y: s.player.y };
-    // Engine-reported speed kept for the log (and the LLM digest)
-    // even though we no longer decide stuck-ness from it.
-    const speed = s.player
-      ? Math.abs(s.player.momx) + Math.abs(s.player.momy)
-      : 1;
-
-    const err =
-      targetBearing !== null && s.player
-        ? angleDelta(targetBearing, s.player.angle_deg)
-        : 0;
-
-    // Forward raycast distance / hit kind: lets us see *why* the
-    // wall-avoid branch fires when reading the trace afterwards.
-    const rays = s.raycasts || [];
-    const fwd = rays.find((r) => Math.abs(r.bearing_deg) < 10);
-    const fwdStr = fwd ? \`\${fwd.hit}@\${fwd.distance.toFixed(0)}\` : "(none)";
-    const enemiesStr =
-      (s.enemies_visible || []).length === 0
-        ? "0"
-        : (s.enemies_visible || [])
-            .map((e) => \`\${e.type}/\${e.bearing}\`)
-            .join(",");
-
-    const branch = await microTick(s, targetBearing, stuckTicks, actuallyMoving);
-    branchCounts[branch.split(" ")[0]] = (branchCounts[branch.split(" ")[0]] || 0) + 1;
-
-    // Log both posDelta (truth) and mom (laggy) so the trace makes
-    // the lag visible: you'll often see early ticks with mom=0
-    // but pdΔ>0 right after a press.
-    await bot.log(
-      \`  t=\${String(t).padStart(2, "0")} \` +
-        \`pos=(\${s.player ? s.player.x.toFixed(0) : "?"},\${s.player ? s.player.y.toFixed(0) : "?"}) \` +
-        \`pdΔ=\${posDelta.toFixed(1)} \` +
-        \`ang=\${s.player ? s.player.angle_deg.toFixed(0) : "?"}° err=\${err.toFixed(0)}° \` +
-        \`mom=(\${s.player ? s.player.momx.toFixed(1) : "?"},\${s.player ? s.player.momy.toFixed(1) : "?"}) \` +
-        \`spd=\${speed.toFixed(1)} stuck=\${stuckTicks} \` +
-        \`fwd=\${fwdStr} enemies=\${enemiesStr} hp=\${s.hud.health} \` +
-        \`-> \${branch}\`,
-    );
-
-    // Reset the counter right after we kicked an unstick so we don't
-    // chain unstick branches forever.
-    if (stuckTicks >= STUCK_TICKS) stuckTicks = 0;
-    await bot.sleep(TICK_MS - 80);
-  }
-
-  // Per-macro summary: how far did we actually move, and which
-  // branches dominated? Also fed back into \`history\` so the next
-  // LLM consult knows whether we got stuck.
-  if (s.player && startPose) {
-    const dx = s.player.x - startPose.x;
-    const dy = s.player.y - startPose.y;
-    const dist = Math.sqrt(dx * dx + dy * dy);
-    const branchStr = Object.entries(branchCounts)
-      .sort((a, b) => b[1] - a[1])
-      .map(([k, v]) => \`\${k}=\${v}\`)
-      .join(" ");
-    const pinned = dist < 32 ||
-      (branchCounts["pinned:"] || 0) + (branchCounts["unstick"] || 0) > 6;
-    await bot.log(
-      \`  macro \${macro} summary: moved \${dist.toFixed(0)} units, branches: \${branchStr}\${pinned ? " [PINNED]" : ""}\`,
-    );
-    history.recent.push({ macro, turn, dist, pinned });
-    if (history.recent.length > 3) history.recent.shift();
-  }
-}
-
-return \`finished \${STEPS} macro steps\`;
-`;
+// Bot example scripts. Each is a real .js file under ./bots/ and
+// is shipped to the codemode sandbox verbatim as a string via
+// Vite's `?raw` import suffix. See AGENTS notes in the README.
+import SIMPLE_BOT from "./bots/simple.js?raw";
+import COMBAT_BOT from "./bots/combat.js?raw";
+import AUTOPLAY_BOT from "./bots/autoplay.js?raw";
+import EXPLORE_BOT from "./bots/explore.js?raw";
+import INSPECT_BOT from "./bots/inspect.js?raw";
+import AI_NAV_BOT from "./bots/ai-nav.js?raw";
 
 interface Example {
 	id: string;
